pyproject.toml
[project]
name = "aion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imap-client>=0.1.8",
    "imap-tools>=1.11.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.10",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
]
